chore(rpc): Moves `SequencerMetrics` into `reth-optimism-rpc` (#18921)
