+++ crates/optimism/rpc/src/lib.rs
@@ -12,6 +12,7 @@ pub mod engine;
 pub mod error;
 pub mod eth;
 pub mod historical;
+pub mod metrics;
 pub mod miner;
 pub mod sequencer;
 pub mod witness;
@@ -21,4 +22,5 @@ pub use engine::OpEngineApiClient;
 pub use engine::{OpEngineApi, OpEngineApiServer, OP_ENGINE_CAPABILITIES};
 pub use error::{OpEthApiError, OpInvalidTransactionError, SequencerClientError};
 pub use eth::{OpEthApi, OpEthApiBuilder, OpReceiptBuilder};
+pub use metrics::SequencerMetrics;
 pub use sequencer::SequencerClient;
