+++ crates/optimism/txpool/src/supervisor/metrics.rs
@@ -1,4 +1,4 @@
-//! Optimism supervisor and sequencer metrics
+//! Optimism supervisor metrics
 
 use crate::supervisor::InteropTxValidatorError;
 use op_alloy_rpc_types::SuperchainDAError;
@@ -68,21 +68,5 @@ impl SupervisorMetrics {
                 _ => {}
             }
         }
-    }
-}
-
-/// Optimism sequencer metrics
-#[derive(Metrics, Clone)]
-#[metrics(scope = "optimism_transaction_pool.sequencer")]
-pub struct SequencerMetrics {
-    /// How long it takes to forward a transaction to the sequencer
-    pub(crate) sequencer_forward_latency: Histogram,
-}
-
-impl SequencerMetrics {
-    /// Records the duration it took to forward a transaction
-    #[inline]
-    pub fn record_forward_latency(&self, duration: Duration) {
-        self.sequencer_forward_latency.record(duration.as_secs_f64());
     }
 }
